requirements.txt
amqp==5.3.1
asgiref==3.8.1
asttokens==3.0.0
billiard==4.2.1
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
decorator==5.1.1
Django==5.1.4
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
executing==2.1.0
inflection==0.5.1
ipython==8.30.0
jedi==0.19.2
kombu==5.4.2
matplotlib-inline==0.1.7
packaging==24.2
parso==0.8.4
prompt_toolkit==3.0.48
psycopg2==2.9.10
pure_eval==0.2.3
Pygments==2.18.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
six==1.17.0
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
tzdata==2024.2
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
